test: remove unreachable worker functional helpers
