+++ docs/internal/baselines/deadcode-baseline.txt
@@ -207,26 +207,4 @@ tests/functional/smoke/service_pipeline_config_helpers_test.go: unreachable func
 tests/functional/transport/acp/stdio/cli_serve_acp_controls_test.go: unreachable func: agentMessageItemIDs
 tests/functional/transport/acp/stdio/cli_serve_acp_prompt_test.go: unreachable func: agentMessageChunkText
 tests/functional/transport/acp/stdio/cli_serve_acp_prompt_test.go: unreachable func: readRPCNotification
-tests/functional/workers/inference/claude/golden_common.go: unreachable func: claudeInvocationResultGolden
-tests/functional/workers/inference/claude/golden_common.go: unreachable func: successfulInferenceResponsePayload
-tests/functional/workers/inference/stream_fidelity_test.go: unreachable func: assertFinalOnlyPublicResponseEvents
-tests/functional/workers/inference/stream_fidelity_test.go: unreachable func: assertSnapshotOnlyPublicResponseEvents
-tests/functional/workers/inference/stream_fidelity_test.go: unreachable func: streamFidelityIdentityResolveSymlinks
-tests/functional/workers/inference/stream_fidelity_test.go: unreachable func: streamFidelityOpenCodeExecutableLocator.LookPath
-tests/functional/workers/inference/stream_fidelity_test.go: unreachable func: writeStreamFidelityOpenCodeExecutable
-tests/functional/workers/mock/live_capacity_test.go: unreachable func: liveCapacityBarrierRunner.callCount
-tests/functional/workers/script/helpers_test.go: unreachable func: assertProviderArgsPrompt
-tests/functional/workers/script/helpers_test.go: unreachable func: assertProviderStdin
-tests/functional/workers/script/helpers_test.go: unreachable func: configureResourceGatedTemplateWorkstation
-tests/functional/workers/script/helpers_test.go: unreachable func: configureTwoInputResourceGatedTemplateWorkstation
-tests/functional/workers/script/helpers_test.go: unreachable func: quoteYAMLString
-tests/functional/workers/script/helpers_test.go: unreachable func: stringPointerValue
-tests/functional/workers/script/helpers_test.go: unreachable func: templateCaptureCommandRunner.LastRequest
-tests/functional/workers/script/helpers_test.go: unreachable func: templateCaptureCommandRunner.Run
-tests/functional/workers/script/helpers_test.go: unreachable func: twoInputTemplateArgs
-tests/functional/workers/script/helpers_test.go: unreachable func: writeFixtureFile
-tests/functional/workers/script/helpers_test.go: unreachable func: writeNamedWorkerAgents
-tests/functional/workers/script/helpers_test.go: unreachable func: writeNamedWorkstationPromptTemplate
-tests/functional/workers/script/helpers_test.go: unreachable func: writeScriptWorkerArgs
-tests/functional/workers/script/helpers_test.go: unreachable func: writeTwoInputResourceSeeds
 tests/stress/guarded_loop_breaker_test.go: unreachable func: assertDispatchHistoryContainsWorkstationRoute
