+++ tests/functional/workers/inference/stream_fidelity_test.go
@@ -1,8 +1,6 @@
 package inference_test
 
 import (
-	"fmt"
-	"os"
 	"path/filepath"
 	"strings"
 	"testing"
@@ -329,171 +327,6 @@ func assertPartialStreamPublicResponseEvents(
 	if !terminalMatched {
 		t.Fatalf("partial-stream replay missing terminal message %q", wantCompletedText)
 	}
-}
-
-func assertSnapshotOnlyPublicResponseEvents(
-	t *testing.T,
-	events []factoryapi.FactoryResponseEvent,
-	wantCompletedText string,
-) {
-	t.Helper()
-
-	var (
-		snapshotCount int
-		terminalSeen  bool
-	)
-
-	for _, event := range events {
-		if event.Kind != factoryapi.FactoryResponseEventKindMessage {
-			continue
-		}
-
-		switch event.Phase {
-		case factoryapi.FactoryResponseEventPhaseDelta:
-			t.Fatalf("snapshot-only replay fabricated message delta: %#v", event)
-
-		case factoryapi.FactoryResponseEventPhaseCompleted:
-			if event.Provenance.Fidelity == factoryapi.FactoryResponseEventProvenanceFidelityFinalOnly {
-				t.Fatalf("snapshot-only replay claimed final-only fidelity on completed message: %#v", event)
-			}
-			if event.Provenance.Delivery != factoryapi.FactoryResponseEventProvenanceDeliveryNativeStream {
-				t.Fatalf(
-					"snapshot-only completed message delivery = %q, want %q: %#v",
-					event.Provenance.Delivery,
-					factoryapi.FactoryResponseEventProvenanceDeliveryNativeStream,
-					event,
-				)
-			}
-			if event.Provenance.Representation != factoryapi.FactoryResponseEventProvenanceRepresentationSnapshot {
-				t.Fatalf(
-					"snapshot-only completed message representation = %q, want %q: %#v",
-					event.Provenance.Representation,
-					factoryapi.FactoryResponseEventProvenanceRepresentationSnapshot,
-					event,
-				)
-			}
-
-			payload, err := event.Payload.AsFactoryResponseEventMessagePayload()
-			if err != nil {
-				t.Fatalf("decode completed message response event: %v", err)
-			}
-			text := streamFidelityMessageText(payload)
-			if text == "" {
-				t.Fatalf("snapshot-only completed message missing text: %#v", event)
-			}
-			snapshotCount++
-			if text == wantCompletedText {
-				terminalSeen = true
-			}
-		}
-	}
-
-	if snapshotCount == 0 {
-		t.Fatal("snapshot-only replay missing completed message snapshots")
-	}
-	if !terminalSeen {
-		t.Fatalf("snapshot-only replay missing terminal completed message %q", wantCompletedText)
-	}
-}
-
-func assertFinalOnlyPublicResponseEvents(
-	t *testing.T,
-	events []factoryapi.FactoryResponseEvent,
-	wantCompletedText string,
-) {
-	t.Helper()
-
-	var (
-		completedCount int
-		terminalSeen   bool
-	)
-
-	for _, event := range events {
-		switch event.Kind {
-		case factoryapi.FactoryResponseEventKindMessage:
-			switch event.Phase {
-			case factoryapi.FactoryResponseEventPhaseDelta:
-				t.Fatalf("final-only replay fabricated message delta: %#v", event)
-
-			case factoryapi.FactoryResponseEventPhaseCompleted:
-				if event.Provenance.Fidelity != factoryapi.FactoryResponseEventProvenanceFidelityFinalOnly {
-					t.Fatalf(
-						"final-only completed message fidelity = %q, want %q: %#v",
-						event.Provenance.Fidelity,
-						factoryapi.FactoryResponseEventProvenanceFidelityFinalOnly,
-						event,
-					)
-				}
-				if event.Provenance.Delivery != factoryapi.FactoryResponseEventProvenanceDeliveryNativeFinal {
-					t.Fatalf(
-						"final-only completed message delivery = %q, want %q: %#v",
-						event.Provenance.Delivery,
-						factoryapi.FactoryResponseEventProvenanceDeliveryNativeFinal,
-						event,
-					)
-				}
-				if event.Provenance.Representation != factoryapi.FactoryResponseEventProvenanceRepresentationSnapshot {
-					t.Fatalf(
-						"final-only completed message representation = %q, want %q: %#v",
-						event.Provenance.Representation,
-						factoryapi.FactoryResponseEventProvenanceRepresentationSnapshot,
-						event,
-					)
-				}
-
-				payload, err := event.Payload.AsFactoryResponseEventMessagePayload()
-				if err != nil {
-					t.Fatalf("decode completed message response event: %v", err)
-				}
-				text := streamFidelityMessageText(payload)
-				if text == "" {
-					t.Fatalf("final-only completed message missing text: %#v", event)
-				}
-				completedCount++
-				if strings.Contains(text, wantCompletedText) {
-					terminalSeen = true
-				}
-			}
-
-		case factoryapi.FactoryResponseEventKindTool:
-			t.Fatalf("final-only replay fabricated tool lifecycle: %#v", event)
-
-		case factoryapi.FactoryResponseEventKindUsage:
-			t.Fatalf("final-only replay fabricated usage lifecycle: %#v", event)
-		}
-	}
-
-	if completedCount == 0 {
-		t.Fatal("final-only replay missing authoritative completed message")
-	}
-	if !terminalSeen {
-		t.Fatalf("final-only replay missing terminal completed message %q", wantCompletedText)
-	}
-}
-
-type streamFidelityOpenCodeExecutableLocator struct {
-	path string
-}
-
-func (l streamFidelityOpenCodeExecutableLocator) LookPath(file string) (string, error) {
-	if file == "opencode" {
-		return l.path, nil
-	}
-	return "", fmt.Errorf("executable %q not found", file)
-}
-
-func streamFidelityIdentityResolveSymlinks(path string) (string, error) {
-	return path, nil
-}
-
-func writeStreamFidelityOpenCodeExecutable(t *testing.T) string {
-	t.Helper()
-
-	path := filepath.Join(t.TempDir(), "opencode")
-	if err := os.WriteFile(path, []byte("opencode-fixture-executable\n"), 0o755); err != nil {
-		t.Fatalf("write opencode fixture executable: %v", err)
-	}
-	return path
 }
 
 func streamFidelityInferenceResponseContains(
